some cleanup for easier reading
